feat: massively expand configuration and refactor core into modular components
