+++ crates/aura-core/Cargo.toml
@@ -27,6 +27,7 @@ suppaftp = { version = "8.0.3", features = ["tokio", "tokio-async-native-tls"] }
 toml = "1.1.2"
 notify = "8.2.0"
 arc-swap = "1.9.1"
+serde_bytes = "0.11.19"
 
 [dev-dependencies]
 tempfile = "3"
